Feat: Chatmate qdrant collection delete #117
