+++ model_hosting/extraction/prompt_extraciont.py
@@ -287,3 +287,20 @@ class PromptExtraction:
     답변:
     {answer}
     """
+
+    def make_prompt_for_civil_complaint(question, context_docs):
+        context = "\n\n".join([f"- {doc}" for doc in context_docs])
+
+        return f"""
+            당신은 민원 응답을 담당하는 공공기관 AI입니다.
+            
+            다음은 국민이 제기한 민원 질문입니다.
+            질문: "{question}"
+            
+            아래는 관련 법령, 과거 Q&A에서 검색된 내용입니다:
+            {context}
+            
+            위 정보를 바탕으로 민원인의 질문에 대해 정확하고 간결하게 답변하십시오.
+            - 법령 조항이 있다면 출처도 함께 제공하십시오.
+            - 친절하고 공손한 말투를 유지하십시오.
+            """
